refactor: modularize terminal components
